requirements.txt
aiogram==2.25.1
requests==2.31.0
python-decouple==3.8
python-dateutil==2.8.2
qrcode==7.4.2
pillow==10.0.1